Extracted Word2VecModelReader to provide a general parser interface
