+++ lucene/analysis/common/src/test/org/apache/lucene/analysis/synonym/TestWord2VecSynonymFilterFactory.java
@@ -17,34 +17,15 @@
 
 package org.apache.lucene.analysis.synonym;
 
-import com.sun.source.tree.AssertTree;
-import org.apache.lucene.analysis.synonym.SynonymProvider.WeightedSynonym;
 import org.apache.lucene.tests.analysis.BaseTokenStreamFactoryTestCase;
 import org.apache.lucene.util.ClasspathResourceLoader;
 import org.apache.lucene.util.ResourceLoader;
-
-import java.io.InputStream;
-import java.util.List;
 
 
 public class TestWord2VecSynonymFilterFactory extends BaseTokenStreamFactoryTestCase {
 
   public static final String FACTORY_NAME = "Word2VecSynonym";
   private static final String WORD2VEC_MODEL_FILE = "word2vec-model.txt";
-
-
-  public void testDl4jModelParser() throws Exception {
-    Word2VecSynonymFilterFactory factory =
-            (Word2VecSynonymFilterFactory) tokenFilterFactory(FACTORY_NAME, "model", WORD2VEC_MODEL_FILE);
-
-    InputStream stream = TestWord2VecSynonymFilterFactory.class.getResourceAsStream(WORD2VEC_MODEL_FILE);
-    List<Word2VecSynonymTerm> terms = factory.loadWordVectors(stream);
-    for (Word2VecSynonymTerm term : terms) {
-      System.out.println(term.getWord());
-    }
-    assertEquals(235, terms.size());
-    assertEquals(100, terms.get(0).getVector().length);
-  }
 
   public void testInform() throws Exception {
     ResourceLoader loader = new ClasspathResourceLoader(getClass());
